requirements.txt
accelerate==0.28.0
datasets==2.18.0
evaluate==0.4.1
tqdm
einops
sentencepiece
scikit-learn
matplotlib
pandas
numpy
ring_flash_attn@git+https://github.com/zhuzilin/ring-flash-attention
deepspeed==0.14.0
transformers==4.39.1
wandb
seaborn
pandas
pytest
termcolor
debugpy
numpy==1.26.4